refactor: Country query tests
