+++ src/Content/Backend/Solution/Monaco.Template.Backend.Application.Tests/Features/Country/GetCountryListTests.cs
@@ -1,25 +1,23 @@
-﻿using FluentAssertions;
+﻿using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using FluentAssertions;
 using Microsoft.Extensions.Primitives;
 using MockQueryable.Moq;
 using Monaco.Template.Backend.Application.DTOs;
 using Monaco.Template.Backend.Application.Features.Country;
 using Monaco.Template.Backend.Application.Infrastructure.Context;
 using Monaco.Template.Backend.Common.Tests.Factories;
-using Monaco.Template.Backend.Common.Tests.Factories.Entities;
 using Moq;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 using Xunit;
 
-namespace Monaco.Template.Backend.Application.Tests.Features.Country.Queries;
+namespace Monaco.Template.Backend.Application.Tests.Features.Country;
 
 [ExcludeFromCodeCoverage]
 [Trait("Application Queries", "Country Queries")]
-public class CountryQueriesHandlersTests
+public class GetCountryListTests
 {
 	[Theory(DisplayName = "Get country list without params succeeds")]
 	[AnonymousData]
@@ -61,34 +59,6 @@ public class CountryQueriesHandlersTests
 			  .HaveCount(countriesSet.Count).And
 			  .Contain(x => countriesSet.Any(c => c.Name == x.Name)).And
 			  .BeInDescendingOrder(x => x.Name);
-	}
-
-	[Fact(DisplayName = "Get existing country by Id succeeds")]
-	public async Task GetExistingCountryByIdSucceeds()
-	{
-		var countries = CountryFactory.CreateMany().ToList();
-		var dbContextMock = SetupMock(countries);
-		var country = countries.First();
-		var query = new GetCountryById.Query(country.Id);
-
-		var sut = new GetCountryById.Handler(dbContextMock.Object);
-		var result = await sut.Handle(query, new CancellationToken());
-
-		result.Should().NotBeNull();
-		result!.Name.Should().Be(country.Name);
-	}
-
-	[Fact(DisplayName = "Get non-existing country by Id fails")]
-	public async Task GetNonExistingCountryByIdFails()
-	{
-		var countries = CountryFactory.CreateMany().ToList();
-		var dbContextMock = SetupMock(countries);
-		var query = new GetCountryById.Query(Guid.NewGuid());
-
-		var sut = new GetCountryById.Handler(dbContextMock.Object);
-		var result = await sut.Handle(query, new CancellationToken());
-
-		result.Should().BeNull();
 	}
 
 	private static Mock<AppDbContext> SetupMock(IEnumerable<Domain.Model.Country> countries)
